% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mailgun.R
\docType{data}
\name{mailgun}
\alias{mailgun}
\title{mailgun factory}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
mailgun
}
\description{
Creates an `IMmailgun` object that allows you to send emails and retrieve stats from your mailgun account.
}
\examples{
\dontrun{
library(IMmailgun)
url <- "https://api.mailgun.net/v3/email.ignacio.website"
key="~/.ssh/id_rsa"
api_key <- secret::get_secret("mailgun", key = key,
vault = file.path(system.file(package = "IMSecrets"),
"vault/"))[[1]]
from="IMSecrets <postmaster@email.ignacio.website>"
email_client <- mailgun$new(url = url, api_key = api_key, from = from)
email_client$sendEmail(to = c("ignacio82@gmail.com", "ignacio@protonmail.com"),
subject = "Test", plaintext = "Hola!")
}
}
\keyword{datasets}
